Rename dataset components
